GitHub/requirements.txt
gradio
openai
python-dotenv
supabase